+++ libsmtutil/SMTPortfolio.cpp
@@ -35,17 +35,19 @@ using namespace solidity::smtutil;
 SMTPortfolio::SMTPortfolio(
 	map<h256, string> _smtlib2Responses,
 	frontend::ReadCallback::Callback _smtCallback,
-	[[maybe_unused]] SMTSolverChoice _enabledSolvers
-)
+	[[maybe_unused]] SMTSolverChoice _enabledSolvers,
+	optional<unsigned> _queryTimeout
+):
+	SolverInterface(_queryTimeout)
 {
-	m_solvers.emplace_back(make_unique<SMTLib2Interface>(move(_smtlib2Responses), move(_smtCallback)));
+	m_solvers.emplace_back(make_unique<SMTLib2Interface>(move(_smtlib2Responses), move(_smtCallback), m_queryTimeout));
 #ifdef HAVE_Z3
 	if (_enabledSolvers.z3)
-		m_solvers.emplace_back(make_unique<Z3Interface>());
+		m_solvers.emplace_back(make_unique<Z3Interface>(m_queryTimeout));
 #endif
 #ifdef HAVE_CVC4
 	if (_enabledSolvers.cvc4)
-		m_solvers.emplace_back(make_unique<CVC4Interface>());
+		m_solvers.emplace_back(make_unique<CVC4Interface>(m_queryTimeout));
 #endif
 }