+++ solc/CommandLineInterface.h
@@ -133,8 +133,8 @@ private:
 	RevertStrings m_revertStrings = RevertStrings::Default;
 	/// Chosen hash method for the bytecode metadata.
 	CompilerStack::MetadataHash m_metadataHash = CompilerStack::MetadataHash::IPFS;
-	/// Chosen model checker engine.
-	ModelCheckerEngine m_modelCheckerEngine = ModelCheckerEngine::All();
+	/// Model checker settings.
+	ModelCheckerSettings m_modelCheckerSettings;
 	/// Whether or not to colorize diagnostics output.
 	bool m_coloredOutput = true;
 	/// Whether or not to output error IDs.